begin modularizing ontology-specific logic
