+++ categorizer/src/main/java/io/dockstore/categorizer/client/cli/Logic.java
@@ -1,0 +1,8 @@
+package io.dockstore.categorizer.client.cli;
+
+import io.dockstore.utils.ai.AIModel;
+import java.util.List;
+
+public interface Logic {
+    List<String> categorize(AIModel aiModel, String entryType, String trsId, String description, String descriptorFileContent);
+}
